requirements.txt
fastapi~=0.112.2
pydantic~=2.10.3
uvicorn~=0.32.1
pyjwt~=2.10.1